hand_gesture/requirements.txt
opencv-python>=4.8.0
cvzone>=1.6.1
mediapipe>=0.10.0
paho-mqtt>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
